pyproject.toml
[project]
name = "issuer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flet>=0.24.0",
    "pillow>=10.4.0",
]

